Re-enable SHA plugin.
The underlying RFC6234 routines are now built using Makedefs.
Fix pointer-to-int-cast warning in SHA plugin (the warning is
raised anywhere where ExtHdrSize (defined in icall.h) is used).
Rename RFC6234/makefile to Makefile.

+++ plugins/src/icall.h
@@ -467,7 +467,7 @@ do {if (sizeof(a[0]) != sizeof(double))  FailCode(102); \
 #define ExternAddr(d) ((void *)&(((struct b_external *)(d).vword.bptr)->exdata[0]))
 
 /* Useful when calling malloc() to get enough space for the block header */
-#define ExtHdrSize  ((int)&(((struct b_external *)(0))->exdata[0]))
+#define ExtHdrSize  ((uintptr_t)&(((struct b_external *)(0))->exdata[0]))
 
 /*
   2. Allocations of Icon structures.
